backend/requirements.txt
fastapi==0.115.0 
uvicorn==0.32.0 
python-multipart==0.0.9 
pillow==10.4.0 
opencv-python==4.10.0.84 
pytesseract==0.3.13 
numpy==1.26.4 
